Make subcommand treatment scalable
